requirements.txt
Flask==2.2.5
python-telegram-bot==13.15
gunicorn==20.1.0